+++ api/src/notifications/notification.service.ts
@@ -29,7 +29,6 @@ import type {
   UpdatePreferencesInput,
 } from './notification.types';
 
-// Re-export types for backward compatibility
 export type {
   ChannelPrefs,
   NotificationType,
@@ -225,8 +224,7 @@ export class NotificationService {
     await this.cronJobService.executeWithTracking(
       'NotificationService_cleanupExpiredNotifications',
       async () => {
-        const deleted = await this.cleanupExpired();
-        if (deleted === 0) return false;
+        if ((await this.cleanupExpired()) === 0) return false;
       },
     );
   }
